update 5 be

+++ app/crud.py
@@ -64,6 +64,8 @@ class UserService:
 
      async def delete_user(self, user_id:int):
           user: User = await self.get_user(user_id=user_id)
+          if not user:
+               raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='User doesnt found!')
           await self.session.delete(user)
 
      @staticmethod
